Split handlers into two files
